Project Added
Added the project.

This project was a graphics engine built on top of the NCLGL framework. Included were implementations of scene Graphs, Mesh and renderer classes and shaders written in GLSL.

Skeletal animations are included as is the water shader which includes dudv mapping, screen space reflections and an implementation of the Fresnel effect.